move out justify code and add tests
